Collapse compliance child-collection differences in the projection pipeline
When EncryptChangeset re-encrypts a read-model snapshot, the comparer reports a
difference for every PII member — including members inside child collections
(contacts.name) — with no array indexers. The MongoDB sink turned that into a
non-positional dotted $set and MongoDB rejected it with WriteError Code 28, so
any projection with a [PII] child collection stopped advancing once a child
existed and an unrelated field changed.

Give the projection path the guarantee the reducer already relied on: collapse
unindexed nested-array differences into a single whole-collection replacement.
Extract that logic into a shared UnindexedCollectionDifferences helper so the
reducer pipeline and EncryptChangeset use one implementation.

A whole-collection $set must also supersede element-level differences on the
same collection, otherwise the collapsed $set conflicts with a positional child
update emitted by the projection in the same changeset (Code 40). Add
GetWholeCollectionReplacementPaths / ConflictsWithWholeCollectionReplacement and
apply the rule across the MongoDB sink, the in-memory sink, and the immediate
projection so element-level differences under a wholesale replacement are dropped.
